LFS-1100: Ugly UI when dragging files over a File question Moved drag and drop to commons and moved styling inside it to be reused in both Variants Upload and File Resource Question.
