+++ modules/data-entry/src/main/frontend/src/questionnaire/QuestionnaireStyle.jsx
@@ -541,20 +541,6 @@ const questionnaireStyle = theme => ({
       borderWidth: '2px',
       borderStyle: 'solid',
     },
-    dropzone: {
-      display: "flex",
-      justifyContent: "center",
-      alignItems: "center",
-      height: theme.spacing(3),
-      width: theme.spacing(44),
-      border: "2px dashed",
-      borderColor: theme.palette.primary.main,
-      padding: "2rem",
-      paddingLeft: "0",
-      textAlign: "center",
-      borderRadius: theme.spacing(1),
-      cursor: "pointer"
-    },
     fileResourceAnswerList: {
       listStyleType: 'none',
       paddingInlineStart: "0",
